+++ src/test/java/kr/co/morandi/backend/domain/contentproblemrecord/random/RandomDefenseProblemRecordTest.java
@@ -43,75 +43,6 @@ class RandomDefenseProblemRecordTest {
         problemRepository.deleteAllInBatch();
         randomDefenseRepository.deleteAllInBatch();
     }
-    @DisplayName("사용자가 랜덤 디펜스를 시작했을 때, 문제 풀이 기록의 정답 여부는 모두 오답이어야 한다.")
-    @Test
-    void isSolvedIsFalse() {
-        // given
-        RandomDefense randomDefense = getRandomDefense();
-        List<Problem> problems = createProblems();
-        Member member = getMember("user");
-        LocalDateTime now = LocalDateTime.now();
-        RandomDefenseRecord randomDefenseRecord
-                = RandomDefenseRecord.create(randomDefense, member, now, problems, randomDefense.getProblemCount());
-        List<ContentProblemRecord> contentProblemRecords = randomDefenseRecord.getContentProblemRecords();
-
-        // when & then
-        assertThat(contentProblemRecords)
-                .extracting("isSolved")
-                .containsExactly(false, false, false, false);
-    }
-
-    @DisplayName("사용자가 랜덤 디펜스를 시작했을 때, 문제 풀이 기록의 제출 횟수는 모두 0회여야 한다.")
-    @Test
-    void submitCountIsZero() {
-        // given
-        RandomDefense randomDefense = getRandomDefense();
-        List<Problem> problems = createProblems();
-        Member member = getMember("user");
-        LocalDateTime now = LocalDateTime.now();
-        RandomDefenseRecord randomDefenseRecord
-                = RandomDefenseRecord.create(randomDefense, member, now, problems, randomDefense.getProblemCount());
-        List<ContentProblemRecord> contentProblemRecords = randomDefenseRecord.getContentProblemRecords();
-
-        // when & then
-        assertThat(contentProblemRecords)
-                .extracting("submitCount")
-                .containsExactly(0L, 0L, 0L, 0L);
-    }
-    @DisplayName("사용자가 랜덤 디펜스를 시작했을 때, 문제 풀이 기록의 정답 코드는 모두 null 이어야 한다.")
-    @Test
-    void solvedCodeIsNull() {
-        // given
-        RandomDefense randomDefense = getRandomDefense();
-        List<Problem> problems = createProblems();
-        Member member = getMember("user");
-        LocalDateTime now = LocalDateTime.now();
-        RandomDefenseRecord randomDefenseRecord
-                = RandomDefenseRecord.create(randomDefense, member, now, problems, randomDefense.getProblemCount());
-        List<ContentProblemRecord> contentProblemRecords = randomDefenseRecord.getContentProblemRecords();
-
-        // when & then
-        assertThat(contentProblemRecords)
-                .extracting("solvedCode")
-                .containsExactly(null, null, null, null);
-    }
-    @DisplayName("사용자가 랜덤 디펜스를 시작했을 때, 문제 풀이 기록의 소요 시간은 모두 0분 이어야 한다.")
-    @Test
-    void solvedTimeIsZero() {
-        // given
-        RandomDefense randomDefense = getRandomDefense();
-        List<Problem> problems = createProblems();
-        Member member = getMember("user");
-        LocalDateTime now = LocalDateTime.now();
-        RandomDefenseRecord randomDefenseRecord
-                = RandomDefenseRecord.create(randomDefense, member, now, problems, randomDefense.getProblemCount());
-        List<ContentProblemRecord> contentProblemRecords = randomDefenseRecord.getContentProblemRecords();
-
-        // when & then
-        assertThat(contentProblemRecords)
-                .extracting("solvedTime")
-                .containsExactly(0L, 0L, 0L, 0L);
-    }
     private static RandomDefense getRandomDefense() {
         RandomCriteria.DifficultyRange bronzeRange = RandomCriteria.DifficultyRange.of(B5, B1);
         RandomCriteria randomCriteria = RandomCriteria.of(bronzeRange, 100L, 200L);
