+++ src/test/java/kr/co/morandi/backend/domain/contentrecord/random/RandomDefenseRecordTest.java
@@ -1,5 +1,6 @@
 package kr.co.morandi.backend.domain.contentrecord.random;
 
+import kr.co.morandi.backend.domain.contentproblemrecord.ContentProblemRecord;
 import kr.co.morandi.backend.domain.contenttype.random.randomcriteria.RandomCriteria;
 import kr.co.morandi.backend.domain.contenttype.random.randomdefense.RandomDefense;
 import kr.co.morandi.backend.domain.contenttype.random.randomdefense.RandomDefenseRepository;
@@ -42,37 +43,47 @@ class RandomDefenseRecordTest {
         problemRepository.deleteAllInBatch();
         randomDefenseRepository.deleteAllInBatch();
     }
-    @DisplayName("사용자가 랜덤 디펜스를 시작했을 때, 맞춘 문제수는 0문제이고 총 문제수는 랜덤 디펜스 문제 개수와 같아야 한다.")
+    @DisplayName("랜덤 디펜스 기록이 만들어졌을 때 맞춘 문제수는 0문제 이어야 한다.")
     @Test
-    void solvedCountIsZeroAndProblemCountEqual() {
+    void solvedCountIsZero() {
         // given
-        RandomCriteria.DifficultyRange bronzeRange = RandomCriteria.DifficultyRange.of(B5, B1);
-        RandomCriteria randomCriteria = RandomCriteria.of(bronzeRange, 100L, 200L);
-        RandomDefense randomDefense = RandomDefense.create(randomCriteria, 4L, 120L, "브론즈 랜덤 디펜스");
+        RandomDefense randomDefense = createRandomDefense();
         List<Problem> problems = createProblems();
-        Member member = makeMember("user");
-        LocalDateTime now = LocalDateTime.now();
+        Member member = createMember("user");
+        LocalDateTime now = LocalDateTime.of(2024, 3, 1, 0,0,0);
 
         // when
         RandomDefenseRecord randomDefenseRecord
                 = RandomDefenseRecord.create(randomDefense, member, now, problems, randomDefense.getProblemCount());
 
         // then
-        assertThat(randomDefenseRecord)
-                .extracting("solvedCount", "problemCount")
-                .containsExactly(0L, randomDefense.getProblemCount());
-
+        assertThat(randomDefenseRecord.getSolvedCount()).isZero();
     }
-    @DisplayName("사용자가 랜덤 디펜스를 시작했을 때, 전체 소요 시간은 0분 이어야 한다.")
+    @DisplayName("랜덤 디펜스 기록이 만들어졌을 때 총 문제 수는 랜덤 디펜스 문제 개수와 같아야 한다.")
+    @Test
+    void problemCountIsEqual() {
+        // given
+        RandomDefense randomDefense = createRandomDefense();
+        List<Problem> problems = createProblems();
+        Member member = createMember("user");
+        LocalDateTime now = LocalDateTime.of(2024, 3, 1, 0,0,0);
+
+        // when
+        RandomDefenseRecord randomDefenseRecord
+                = RandomDefenseRecord.create(randomDefense, member, now, problems, randomDefense.getProblemCount());
+
+        // then
+        assertThat(randomDefenseRecord.getProblemCount()).isEqualTo(randomDefense.getProblemCount());
+    }
+
+    @DisplayName("랜덤 디펜스 기록이 만들어졌을 때 전체 소요 시간은 0분 이어야 한다.")
     @Test
     void totalSolvedTimeIsZero() {
         // given
-        RandomCriteria.DifficultyRange bronzeRange = RandomCriteria.DifficultyRange.of(B5, B1);
-        RandomCriteria randomCriteria = RandomCriteria.of(bronzeRange, 100L, 200L);
-        RandomDefense randomDefense = RandomDefense.create(randomCriteria, 4L, 120L, "브론즈 랜덤 디펜스");
+        RandomDefense randomDefense = createRandomDefense();
         List<Problem> problems = createProblems();
-        Member member = makeMember("user");
-        LocalDateTime now = LocalDateTime.now();
+        Member member = createMember("user");
+        LocalDateTime now = LocalDateTime.of(2024, 3, 1, 0,0,0);
 
         // when
         RandomDefenseRecord randomDefenseRecord
@@ -81,16 +92,14 @@ class RandomDefenseRecordTest {
         // then
         assertThat(randomDefenseRecord.getTotalSolvedTime()).isZero();
     }
-    @DisplayName("사용자가 랜덤 디펜스를 시작한 시점과 랜덤 디펜스 시험 날짜는 같아야 한다.")
+    @DisplayName("랜덤 디펜스 기록이 만들어졌을 때 시점과 랜덤 디펜스 시험 날짜는 같아야 한다.")
     @Test
     void testDateIsEqualNow() {
         // given
-        RandomCriteria.DifficultyRange bronzeRange = RandomCriteria.DifficultyRange.of(B5, B1);
-        RandomCriteria randomCriteria = RandomCriteria.of(bronzeRange, 100L, 200L);
-        RandomDefense randomDefense = RandomDefense.create(randomCriteria, 4L, 120L, "브론즈 랜덤 디펜스");
+        RandomDefense randomDefense = createRandomDefense();
         List<Problem> problems = createProblems();
-        Member member = makeMember("user");
-        LocalDateTime now = LocalDateTime.now();
+        Member member = createMember("user");
+        LocalDateTime now = LocalDateTime.of(2024, 3, 1, 0,0,0);
 
         // when
         RandomDefenseRecord randomDefenseRecord
@@ -99,9 +108,90 @@ class RandomDefenseRecordTest {
         // then
         assertThat(randomDefenseRecord.getTestDate()).isEqualTo(now);
     }
+    @DisplayName("랜덤 디펜스 기록이 만들어졌을 때 세부 문제 기록의 정답 여부는 모두 오답이어야 한다.")
+    @Test
+    void isSolvedIsFalse() {
+        // given
+        RandomDefense randomDefense = createRandomDefense();
+        List<Problem> problems = createProblems();
+        Member member = createMember("user");
+        LocalDateTime now = LocalDateTime.of(2024, 3, 1, 0,0,0);
 
+        // when
+        RandomDefenseRecord randomDefenseRecord
+                = RandomDefenseRecord.create(randomDefense, member, now, problems, randomDefense.getProblemCount());
+        List<ContentProblemRecord> contentProblemRecords = randomDefenseRecord.getContentProblemRecords();
 
-    private Member makeMember(String name) {
+        // then
+        assertThat(contentProblemRecords)
+                .extracting("isSolved")
+                .containsExactly(false, false, false, false);
+    }
+
+    @DisplayName("랜덤 디펜스 기록이 만들어졌을 때 세부 문제 기록의 제출 횟수는 모두 0회여야 한다.")
+    @Test
+    void submitCountIsZero() {
+        // given
+        RandomDefense randomDefense = createRandomDefense();
+        List<Problem> problems = createProblems();
+        Member member = createMember("user");
+        LocalDateTime now = LocalDateTime.of(2024, 3, 1, 0,0,0);
+
+        // when
+        RandomDefenseRecord randomDefenseRecord
+                = RandomDefenseRecord.create(randomDefense, member, now, problems, randomDefense.getProblemCount());
+        List<ContentProblemRecord> contentProblemRecords = randomDefenseRecord.getContentProblemRecords();
+
+        // then
+        assertThat(contentProblemRecords)
+                .extracting("submitCount")
+                .containsExactly(0L, 0L, 0L, 0L);
+    }
+    @DisplayName("랜덤 디펜스 기록이 만들어졌을 때 세부 문제 기록의 정답 코드는 모두 null 이어야 한다.")
+    @Test
+    void solvedCodeIsNull() {
+        // given
+        RandomDefense randomDefense = createRandomDefense();
+        List<Problem> problems = createProblems();
+        Member member = createMember("user");
+        LocalDateTime now = LocalDateTime.of(2024, 3, 1, 0,0,0);
+
+        // when
+        RandomDefenseRecord randomDefenseRecord
+                = RandomDefenseRecord.create(randomDefense, member, now, problems, randomDefense.getProblemCount());
+        List<ContentProblemRecord> contentProblemRecords = randomDefenseRecord.getContentProblemRecords();
+
+        // then
+        assertThat(contentProblemRecords)
+                .extracting("solvedCode")
+                .containsExactly(null, null, null, null);
+    }
+    @DisplayName("랜덤 디펜스 기록이 만들어졌을 때 세부 문제 기록의 소요 시간은 모두 0분 이어야 한다.")
+    @Test
+    void solvedTimeIsZero() {
+        // given
+        RandomDefense randomDefense = createRandomDefense();
+        List<Problem> problems = createProblems();
+        Member member = createMember("user");
+        LocalDateTime now = LocalDateTime.of(2024, 3, 1, 0,0,0);
+
+        // when
+        RandomDefenseRecord randomDefenseRecord
+                = RandomDefenseRecord.create(randomDefense, member, now, problems, randomDefense.getProblemCount());
+        List<ContentProblemRecord> contentProblemRecords = randomDefenseRecord.getContentProblemRecords();
+
+        // then
+        assertThat(contentProblemRecords)
+                .extracting("solvedTime")
+                .containsExactly(0L, 0L, 0L, 0L);
+    }
+    private RandomDefense createRandomDefense() {
+        RandomCriteria.DifficultyRange bronzeRange = RandomCriteria.DifficultyRange.of(B5, B1);
+        RandomCriteria randomCriteria = RandomCriteria.of(bronzeRange, 100L, 200L);
+        RandomDefense randomDefense = RandomDefense.create(randomCriteria, 4L, 120L, "브론즈 랜덤 디펜스");
+        return randomDefense;
+    }
+    private Member createMember(String name) {
         Member member = Member.create(name, name + "@gmail.com", GOOGLE, name, name);
         return memberRepository.save(member);
     }
